+++ rust/src/producers/mod.rs
@@ -1,5 +1,7 @@
 /// Example statements generator.
 pub mod examples;
+/// Boolean example statements generator.
+pub mod boolean_examples;
 /// A Sink stores produced messages, e.g. in files.
 pub mod sink;